fix categoria test

+++ microservice-categoria/src/main/java/com/microservice/categoria/domain/repositories/CategoriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Integer> {
+public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Long> {
 }
