+++ microservice-categoria/src/main/java/com/microservice/categoria/infraestructure/abstract_service/ICategoriaService.java
@@ -1,7 +1,5 @@
 package com.microservice.categoria.infraestructure.abstract_service;
 
-import com.microservice.categoria.api.modeles.request.CategoriaRequest;
-import com.microservice.categoria.api.modeles.responses.CategoriaResponse;
 import com.microservice.categoria.domain.entities.CategoriaEntity;
 
 import java.util.List;
@@ -9,8 +7,12 @@ import java.util.List;
 public interface ICategoriaService {
 
     CategoriaEntity save(CategoriaEntity categoria);
-    CategoriaEntity update(Integer id, CategoriaEntity categoria);
-    boolean delete(Integer id);
-    CategoriaEntity findById(Integer id);
+
+    CategoriaEntity update(Long id, CategoriaEntity categoria);
+
+    boolean delete(Long id);
+
+    CategoriaEntity findById(Long id);
+
     List<CategoriaEntity> findAll();
 }
